+++ cli/cli/commands/exec/exec.go
@@ -39,6 +39,8 @@ const (
 
 	scriptArgForLogging = "script"
 	moduleArgForLogging = "module"
+
+	isNewEnclave = true
 )
 
 var StartosisExecCmd = &lowlevel.LowlevelKurtosisCommand{
@@ -141,11 +143,13 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred connecting to the local Kurtosis engine")
 	}
 
-	enclaveCtx, err := getOrCreateEnclaveContext(ctx, enclaveId, kurtosisCtx, isPartitioningEnabled)
+	enclaveCtx, isNewEnclave, err := getOrCreateEnclaveContext(ctx, enclaveId, kurtosisCtx, isPartitioningEnabled)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the enclave context for enclave '%v'", enclaveId)
 	}
-	defer output_printers.PrintEnclaveId(enclaveCtx.GetEnclaveID())
+	if isNewEnclave {
+		defer output_printers.PrintEnclaveId(enclaveCtx.GetEnclaveID())
+	}
 
 	fileOrDir, err := os.Stat(startosisScriptOrModulePath)
 	if err != nil {
@@ -255,23 +259,31 @@ func validateExecutionResponse(executionResponse *kurtosis_core_rpc_api_bindings
 	return nil
 }
 
-func getOrCreateEnclaveContext(ctx context.Context, enclaveId enclaves.EnclaveID, kurtosisContext *kurtosis_context.KurtosisContext, isPartitioningEnabled bool) (*enclaves.EnclaveContext, error) {
+func getOrCreateEnclaveContext(
+	ctx context.Context,
+	enclaveId enclaves.EnclaveID,
+	kurtosisContext *kurtosis_context.KurtosisContext,
+	isPartitioningEnabled bool,
+) (*enclaves.EnclaveContext, bool, error) {
+
 	enclavesMap, err := kurtosisContext.GetEnclaves(ctx)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "Unable to get existing enclaves from Kurtosis backend")
+		return nil, false, stacktrace.Propagate(err, "Unable to get existing enclaves from Kurtosis backend")
 	}
 	if _, found := enclavesMap[enclaveId]; found {
 		enclaveContext, err := kurtosisContext.GetEnclaveContext(ctx, enclaveId)
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "Unable to get enclave context from the existing enclave '%s'", enclaveId)
+			return nil, false, stacktrace.Propagate(err, "Unable to get enclave context from the existing enclave '%s'", enclaveId)
 		}
-		return enclaveContext, nil
+		return enclaveContext, false, nil
 	}
+	logrus.Infof("Creating a new enclave for the startosis script to execute inside...")
 	enclaveContext, err := kurtosisContext.CreateEnclave(ctx, enclaveId, isPartitioningEnabled)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, fmt.Sprintf("Unable to create new enclave with ID '%s'", enclaveId))
+		return nil, false, stacktrace.Propagate(err, fmt.Sprintf("Unable to create new enclave with ID '%s'", enclaveId))
 	}
-	return enclaveContext, nil
+	logrus.Infof("Enclave '%v' created successfully", enclaveContext.GetEnclaveID())
+	return enclaveContext, isNewEnclave, nil
 }
 
 // validateModuleArgs just validates the args is a valid JSON string