+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/wait/wait.go
@@ -23,13 +23,14 @@ const (
 func GenerateWaitBuiltin(instructionsQueue *[]kurtosis_instruction.KurtosisInstruction, factsEngine *facts_engine.FactsEngine) func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	// TODO: Force returning an InterpretationError rather than a normal error
 	return func(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
-		serviceId, commandArgs, interpretationError := parseStartosisArgs(b, args, kwargs)
+		serviceId, factName, interpretationError := parseStartosisArgs(b, args, kwargs)
 		if interpretationError != nil {
 			return nil, interpretationError
 		}
-		waitInstruction := NewWaitInstruction(factsEngine, *shared_helpers.GetCallerPositionFromThread(thread), serviceId, commandArgs)
+		waitInstruction := NewWaitInstruction(factsEngine, *shared_helpers.GetCallerPositionFromThread(thread), serviceId, factName)
 		*instructionsQueue = append(*instructionsQueue, waitInstruction)
-		return starlark.None, nil
+		returnValue := shared_helpers.MakeWaitInterpretationReturnValue(serviceId, factName)
+		return returnValue, nil
 	}
 }