+++ internal_testsuites/golang/testsuite/facts_engine_test/facts_engine_test.go
@@ -84,7 +84,10 @@ func TestFactsEngine(t *testing.T) {
 	require.Equal(t, expectedOutputForExecFactOutput, getFactValuesResponse.GetFactValues()[0].GetStringValue())
 
 	logrus.Infof("Defining HTTP request fact...")
-	_, err = enclaveCtx.DefineFact(binding_constructors.NewGetHttpRequestFactRecipeWithDefaultRefresh(testServiceId, httpRequestFactName, containerPortId, "/"))
+	_, err = enclaveCtx.DefineFact(&kurtosis_core_rpc_api_bindings.FactRecipe{
+		ServiceId:            testServiceId,
+		FactName:             httpRequestFactName,
+		FactRecipeDefinition: binding_constructors.NewGetHttpRequestFactRecipeDefinition(containerPortId, "/", nil)})
 	require.Nil(t, err)
 	time.Sleep(5 * time.Second)